feat: Add DTOs for active projects, time logs, and timer requests
- Created ActiveProjectDTO for managing active project data.
- Introduced CreateTimeLogRequestDTO for creating time logs with validation.
- Added StartTimerRequestDTO and StopTimerRequestDTO for timer operations.
- Implemented TimeLogResponseDTO and TimeLogSummaryDTO for response structures.
- Developed TimerResponseDTO for timer-related responses.
- Added UpdateTimeLogRequestDTO for updating existing time logs.

feat: Implement entity classes for appointments, employees, time logs, and timers

- Created Appointment entity with relationships to User and Employee.
- Developed Employee entity with time logs and appointments relationships.
- Implemented TimeLog entity with references to Employee and Appointment.
- Added Timer entity to manage active timers for appointments.

feat: Create repositories for data access

- Added AppointmentRepository for appointment-related database operations.
- Implemented EmployeeRepository for managing employee data.
- Created TimeLogRepository for time log operations.
- Developed TimerRepository for timer management.

chore: Update application properties for database migration

- Changed Hibernate DDL auto setting to update for schema management.
- Configured Flyway for database migrations with baseline settings.

chore: Add SQL migration script for creating necessary tables

- Created V3__create_appointments_and_time_logs.sql for initializing employees, appointments, and time logs tables.
