+++ src/main/java/com/autocare360/dto/ActiveProjectDTO.java
@@ -1,4 +1,4 @@
-package com.example.autocare360.dto;
+package com.autocare360.dto;
 
 public class ActiveProjectDTO {
