+++ src/main/java/com/autocare360/dto/StartTimerRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.autocare360.dto;
+package com.autocare360.dto;
 
 import jakarta.validation.constraints.NotNull;
