+++ src/main/java/com/autocare360/repo/AppointmentRepository.java
@@ -1,0 +1,13 @@
+package com.autocare360.repo;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.autocare360.entity.Appointment;
+
+@Repository
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+    List<Appointment> findByStatusOrderByDateAscTimeAsc(String status);
+}
