+++ src/main/java/com/autocare360/repo/EmployeeRepository.java
@@ -1,11 +1,11 @@
-package com.example.autocare360.repository;
+package com.autocare360.repo;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.autocare360.entity.Employee;
+import com.autocare360.entity.Employee;
 
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
