+++ src/main/java/com/autocare360/repo/TimerRepository.java
@@ -1,11 +1,11 @@
-package com.example.autocare360.repository;
+package com.autocare360.repo;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.autocare360.entity.Timer;
+import com.autocare360.entity.Timer;
 
 @Repository
 public interface TimerRepository extends JpaRepository<Timer, Long> {
