+++ src/main/resources/application.properties.example
@@ -5,13 +5,16 @@ spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.jdbc.time_zone=UTC
 
 spring.flyway.enabled=true
+spring.flyway.baseline-version=3
+spring.flyway.baseline-on-migrate=true
+spring.flyway.validate-on-migrate=false
 
 # JWT & CORS
 app.security.jwt.secret=${JWT_SECRET}
